Move filter to a new pkg
