+++ pkg/filter/circularBuffer_test.go
@@ -1,4 +1,4 @@
-package honeycomb
+package filter
 
 import (
 	"fmt"
