+++ pkg/filter/filter.go
@@ -1,4 +1,4 @@
-package honeycomb
+package filter
 
 import (
 	"bufio"
