+++ pkg/filter/jsonsplit.go
@@ -1,4 +1,4 @@
-package honeycomb
+package filter
 
 import (
 	"bytes"
